from double to long double, setprecision stream, clean PDE solver ...

+++ include/Ecoli/CV_rExpIG_tExp.h
@@ -29,7 +29,7 @@ public:
     
     void print_info(std::ofstream &file_save);
     void reset_par();
-    double reset_barrier();
+    long double reset_barrier();
 
 };
