+++ include/Global/GlobalVariables.h
@@ -9,7 +9,7 @@
 #include <vector>
 #include <mutex>
 
-extern std::vector<double> vettore_risposte;
+extern std::vector<long double> vettore_risposte;
 extern std::vector<std::string> vettore_text;
 extern int contatore_risposte;
 
@@ -27,8 +27,8 @@ extern bool colors;
 /** 
  * Variable to keep track of the MODE of simulation ... in otherwords, how to get the imput,
  * from keybord and run the imput (0)
- * from keybord and save the imput in std::vector<double> vettore_risposte; contatore_risposte keep track of the answer we read from the std::vector of answers (1)
- * read from std::vector<double> vettore_risposte (2)
+ * from keybord and save the imput in std::vector<long double> vettore_risposte; contatore_risposte keep track of the answer we read from the std::vector of answers (1)
+ * read from std::vector<long double> vettore_risposte (2)
  automatic: 0 -> step by step version
             1 -> setting parameters
             2 -> running simulation
@@ -36,7 +36,7 @@ extern bool colors;
 extern int automatic_;
 
 /** variable used to study the Diffusive limit*/
-extern double epsilon;
+extern long double epsilon;
 
 /** Variable used to reseed the random generator to allow the possibility to use the same random numbers in one Routine */
 extern unsigned int seed_;
